feat: RouteService 클래스 생성 - route.js 삭제 - RouteService 클래스는 싱글톤 패턴으로 구현
